+++ src/main/scala/com/itszuvalex/itszucore/implicits/package.scala
