+++ src/main/scala/com/itszuvalex/itszucore/util/PlayerUtils.scala
@@ -18,7 +18,7 @@
  *  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  *  *****************************************************************************
  */
-package com.itszuvalex.itszucore.api.util
+package com.itszuvalex.itszucore.util
 
 import java.util.UUID
